mvp api functional after refactor, nswag
added migration to prod env, api client generation with nswag, api is functional and refactored ! (not fully tested yet)

+++ Explore.Domain/TenantUser.cs
@@ -15,7 +15,7 @@ namespace Explore.Domain
         public Guid TenantId { get; set; }
         public Tenant Tenant { get; set; }
         [ForeignKey("UserRole")]
-        public long UserRoleId { get; set; }
+        public int UserRoleId { get; set; }
         public UserRole UserRole { get; set; }
     }
 }
